[Fix] 버그 수정
1. 공격력이 너무 높으면 적의 체력이 증가하는 현상 수정
2. 적이 일정 스테이지 이후로 공격하지 않는 현상 수정

+++ Growing-a-Cactus/Assets/Scripts/Enemy&Boss/EnemyScript.cs
@@ -67,7 +67,7 @@ public class EnemyScript : MonoBehaviour
 
     public void TakeDamage(double damage)
     {
-        HP -= (int)damage;
+        HP -= damage;
 
         UpdateHPBar();
